refactor(ai): extract KB-resolution cluster into shotsummarizer_kb.cpp (no behavior change)
The profile-knowledge resolution layer (loadProfileKnowledge,
matchProfileKey, computeProfileKbId, getAnalysisFlags, ugs*,
canonicalNameForKbId, expertBandForKbId, buildProfileCatalog,
crossProfileReferenceContent, allKbUgsEntries + the s_profileKnowledge/
catalog/dial-in/cross-profile statics + normalizeProfileKey) was welded
into the monolithic shotsummarizer.cpp, forcing any offline consumer
(shot_eval, tst_shotrecord_cache) to drag the whole app
(shotdatamodel/visualizeruploader[Qt Network]/dialing) in.

Moved verbatim (src/ai/shotsummarizer.cpp:1374-1770 + statics 24-51)
into a new lean TU src/ai/shotsummarizer_kb.cpp depending only on the
:/ai/*.md resources + QString + ShotAnalysis::ExpertBand. Declarations
stay in shotsummarizer.h as ShotSummarizer:: statics — zero API/header/
caller change; the class-static definitions live once in the new TU,
referenced by both TUs. Added to the app + the four test targets that
list shotsummarizer.cpp.

Enables shot_eval to link a small closure and route through the SAME
prepareAnalysisInputs->analyzeShot->deriveBadges path as the app
(parity by construction — next commit). Pure structural split: build
0/0, full suite 2083/0/0 (behavior-preserving by construction).
